Created process_data3.py
Improvement on the process_data2 file that allows the simultaneous calculations of the means of multiple variables of an LDAS file type and writes them to the same CSV file. It has an unused function to add new data to a previously created CSV file. It requires configuring the file locations, including the output CSV and list of variables to compute the means.