Create tseries_point_plot.py
Code to generate a scatter plot time series graph from a CSV file. The x-axis is the time gotten from the time column of the file, and the y-axis is the individual divide catchments from the divide_id column. The variable magnitude is shown with a color bar that also corresponds to the z-axis in the 3d plot.